Fix sound manager when not supported
The sound manager checked for support but then assumed support. This was
an oversight not obvious at the time when chrome supported the Webkit
Audio Context. Now that chrome has moved on it's broken the game. This
commit mocks problematic methods on the sound manager with no-ops so
that the game can be played without sound.